Changes in this commit
* Add matrices
* Minor changes in vectors
* Rename PipelineLayout class name with LayoutManager
* Rename PixelBufferAccess class name with BufferTexelAccess
* Add C++ templace utilities and helper functions
* Add index for shaders to ProgramCollection class
* Add ZSubpassDependency to barriers
* Add dynamic rendering functionality
* Add multiview renderpass
* Add multiview tests
* Add int_matrix tests
* Add topology tests
* Add line_width tests
* Upgrade some pragmas

+++ framework/vtfBacktrace.hpp
@@ -14,7 +14,7 @@ template<size_t Depth>
 void printBacktrace (std::ostream& ss, std::size_t skip = 0)
 {
 	void* bts[Depth];
-	const int	nbt = backtrace(bts, Depth);
+	const uint32_t	nbt = static_cast<uint32_t>(backtrace(bts, Depth));
 	printBts(ss, bts, nbt, skip);
 }
